+++ library/e_terminal.h
@@ -42,8 +42,8 @@ struct _e_terminal : public _ui_element
 	void start_timer();
 	void stop_timer(std::wstring_view s);
 	void update() override;
-	void key_down(ushort key) override;
-	void key_press(ushort key) override;
+	void key_down(u64 key) override;
+	void key_press(u64 key) override;
 	bool mouse_wheel2(_xy r, short value) override;
 	bool mouse_down_left2(_xy r) override;
 	void mouse_move_left2(_xy r) override;
